Terminado el TP4 y testeado

+++ trabajo_practico_4/likedList_Eclipse_v9/likedList/testing/inc/Employee.h
@@ -22,7 +22,7 @@ struct
     int id;
     char name[51];
     char lastName[51];
-    float salary;
+    int salary;
     int sector;
     int isEmpty;
